requirements-dev.txt
-r requirements.txt
django-extensions>=3.1,<3.2
django-debug-toolbar>=2.2,<2.3
Fabric>=2.6,<2.7
flake8>=4,<4.1
ipython>=7.16,<7.17
tox>=3.24,<3.25
